refactor: util 패키지 경로 이동
