+++ src/main/java/attendance/controller/util/CrewAttendancesDataParser.java
@@ -1,4 +1,4 @@
-package attendance.util;
+package attendance.controller.util;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
