+++ src/test/java/attendance/domain/AttendanceBookFactoryTest.java
@@ -1,12 +1,11 @@
 package attendance.domain;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-import attendance.util.AttendanceBookFactory;
+import attendance.controller.util.AttendanceBookFactory;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
+import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.Test;
 
 class AttendanceBookFactoryTest {
